% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_supabase.R
\name{obtener_datos_lublin_esp}
\alias{obtener_datos_lublin_esp}
\title{Obtener datos de españoles en Lublin (ID 38)}
\usage{
obtener_datos_lublin_esp()
}
\description{
Obtener datos de españoles en Lublin (ID 38)
}
