% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_supabase.R
\name{obtener_marcas_ciclo_olimpico}
\alias{obtener_marcas_ciclo_olimpico}
\title{Obtener Top 50 marcas por año para el Ciclo Olímpico}
\usage{
obtener_marcas_ciclo_olimpico(distancia_in, estilo_in, genero_in)
}
\arguments{
\item{distancia_in}{Distancia de la prueba (ej: 100)}

\item{estilo_in}{Estilo de la prueba (ej: 'FREE')}

\item{genero_in}{Género (ej: 'M' o 'F')}
}
\description{
Obtener Top 50 marcas por año para el Ciclo Olímpico
}
